Fix unit tests after removing engine module
