+++ Net SDK Unit Tests/Unit Tests/Client/LocalhostClientUnitTests.cs
@@ -1,0 +1,43 @@
+﻿using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Splitio.Services.Client.Classes;
+using Splitio.Services.EngineEvaluator;
+using Moq;
+using System.Collections.Generic;
+using Splitio.Domain;
+
+namespace Splitio_Tests.Unit_Tests.Client
+{
+    [TestClass]
+    public class LocalhostClientUnitTests
+    {
+        
+        [TestMethod]
+        [DeploymentItem(@"Resources\test.splits")]
+        public void GetTreatmentShouldReturnControlIfSplitNotFound()
+        {
+            //Arrange
+            var splitClient = new LocalhostClient("test.splits");
+            
+            //Act
+            var result = splitClient.GetTreatment("test", "test");
+
+            //Assert
+            Assert.AreEqual("control", result);
+        }
+        
+        [TestMethod]
+        [DeploymentItem(@"Resources\test.splits")]
+        public void GetTreatmentShouldRunAsSingleKeyUsingNullBucketingKey()
+        {
+            var splitClient = new LocalhostClient("test.splits");
+
+            //Act
+            var key = new Key("test", null);
+            var result = splitClient.GetTreatment(key, "other_test_feature");
+
+            //Assert
+            Assert.AreEqual(key.bucketingKey, key.matchingKey);
+        }
+    }
+}
